+++ conf/tests/test_sitemaps.py
@@ -2,7 +2,7 @@ import http
 
 import pytest
 
-from django.core.urlresolvers import reverse
+from django.urls import reverse
 
 
 @pytest.mark.django_db