+++ sso/user/views.py
@@ -2,7 +2,7 @@ import urllib
 
 from django.db import IntegrityError
 from django.conf import settings
-from django.core.urlresolvers import reverse
+from django.urls import reverse
 from django.shortcuts import redirect
 from django.views.generic import RedirectView
 from django.urls import reverse_lazy
@@ -42,13 +42,9 @@ class SignupView(DisableRegistrationMixin, RedirectToNextMixin, allauth_views.Si
 
     @staticmethod
     def is_email_not_unique_error(integrity_error):
-        email_not_unique_message = (
-            'duplicate key value violates unique '
-            'constraint "user_user_email_key"'
-        )
-        return any((
-            email_not_unique_message in arg for arg in integrity_error.args
-        ))
+        email_not_unique_message = 'duplicate key value violates unique constraint "user_user_email_key"'
+
+        return any((email_not_unique_message in arg for arg in integrity_error.args))
 
     def form_valid(self, form):
         try:
@@ -76,7 +72,7 @@ class LoginView(RedirectToNextMixin, allauth_views.LoginView):
         response = super().form_valid(form)
         if response.status_code == 302 and response.url == reverse("account_email_verification_sent"):
             return response
-        elif settings.FEATURE_FLAGS['NEW_ENROLMENT_ON']:
+        else:
             self.request.session.save()
             has_profile = utils.user_has_profile(form.user)
             if not has_profile:
@@ -140,3 +136,13 @@ class PasswordChangeView(DisableRegistrationMixin,
 
 class SSOLandingPage(RedirectView):
     url = reverse_lazy('account_login')
+
+
+class LoginViaLinkedinView(RedirectView):
+    def get_redirect_url(self):
+        return utils.get_login_provider_url(request=self.request, provider_id='linkedin_oauth2')
+
+
+class LoginViaGoogleView(RedirectView):
+    def get_redirect_url(self):
+        return utils.get_login_provider_url(request=self.request, provider_id='google')